Added demo game/asteroids (#1)
* Initial project setup

- Removed old assets
- Created Asteroids project

* Project settings tweaks

- Added title/main scenes

* Added game icon

* Organized main.lua

* Added font

* Added Asteroids logo texture

- Also working on title scene

* Added music files

- Imported music assets
- Set default music asset for each scene

* Fixed UI spacing in content browser

* Added texture assets to the project

* Added background to main scene

* Added sound fx assets to the project

* Added layout of objects to main scene

* Added player prefab

* Updated object data

* Added lua scripts

* Updated API in SpriteComponent

- hidden -> isHidden

* Fixed bug with collisions

* Fixed sprites

* Fixed gamepad inputs

* Fixed pickup sprites

* Added ability to play soundfx/music directly from asset context popup

* Fixed "exclude" lua binding

* Small fixes

* Fixed game object duplication

- The ability to duplicate game objects did not copy relationships correctly
- Added new helper functions DuplicateRecursive and ApplyFunctionToHierarchy
- DuplicateRecursive will successfully copy all entities and relationships
- ApplyfunctionToHierarchy will apply a passed in callable on the source entity and all of its children etc
- Added check to ensure we cannot add uneditable entities as child entities or add child entities to uneditable.
- Adjusted Editor core wrappers bindings to use the same naming conventions as the core

* Defer entity destruction

* Fixed StateMachine::ChangeState

* Refactor

* Added descriptions to Open Dialog

* Revert

+++ Feather-Core/src/Utils/Tween.cpp
@@ -116,7 +116,7 @@ namespace Feather {
 		float a = change;
 		float s = p / 4.0f;
 
-		return -(a * powf(2.0f, -10 * (current -= 1)) * sin((current * duration - s) * TwoPI / p)) + start;
+		return -(a * powf(2.0f, -10 * (current -= 1.0f)) * sinf((current * duration - s) * TwoPI / p)) + start;
 	};
 
 	// Elastic easing: [OUT] - Oscillates around the end position.
@@ -132,7 +132,7 @@ namespace Feather {
 		float a = change;
 		float s = p / 4.0f;
 
-		return -(a * powf(2.0f, -10 * current) * sin((current * duration - s) * TwoPI / p)) + change + start;
+		return -(a * powf(2.0f, -10.0f * current) * sinf((current * duration - s) * TwoPI / p)) + change + start;
 	};
 
 	// Elastic easing: [IN-OUT] - Oscillates around the end position.
@@ -149,9 +149,9 @@ namespace Feather {
 		float s = p / 4.0f;
 
 		if (current < 1.0f)
-			return -0.5f * (a * powf(2.0f, -10 * (current -= 1)) * sin((current * duration - s) * TwoPI / p)) + start;
+			return -0.5f * (a * powf(2.0f, -10.0f * (current -= 1.0f)) * sinf((current * duration - s) * TwoPI / p)) + start;
 
-		return (a * powf(2.0f, -10 * (current -= 1.0f)) * sin((current * duration - s) * TwoPI / p)) * 0.5f + change + start;
+		return (a * powf(2.0f, -10.0f * (current -= 1.0f)) * sinf((current * duration - s) * TwoPI / p)) * 0.5f + change + start;
 	};
 
 	// Exponential easing: 2^t
